pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "admin-chat-bot"
version = "1.0.0"
description = "Telegram-бот для административных чатов с ролевой системой"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Ваше Имя", email = "your.email@example.com"}
]
dependencies = [
    "aiogram>=3.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.2.1",
    "pytest-asyncio>=0.20.3",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["app", "handlers", "keyboards", "models", "utils"] 